requirements_1767202417847.txt
Flask>=2.3.2
discord.py>=2.3.0
